+++ GloryECS/GloryECSStatic/EntityView.h
@@ -14,10 +14,10 @@ namespace GloryECS
 
 		size_t ComponentCount();
 		Glory::UUID ComponentUUIDAt(size_t index);
-		size_t ComponentTypeAt(size_t index);
-		size_t ComponentType(Glory::UUID uuid);
-		std::map<Glory::UUID, size_t>::iterator GetIterator();
-		std::map<Glory::UUID, size_t>::iterator GetIteratorEnd();
+		uint32_t ComponentTypeAt(size_t index);
+		uint32_t ComponentType(Glory::UUID uuid);
+		std::map<Glory::UUID, uint32_t>::iterator GetIterator();
+		std::map<Glory::UUID, uint32_t>::iterator GetIteratorEnd();
 		void SwapComponentIndex(size_t index1, size_t index2);
 		void SetComponentIndex(size_t oldIndex, size_t newIndex);
 
@@ -29,8 +29,8 @@ namespace GloryECS
 	private:
 		friend class EntityRegistry;
 		EntityRegistry* m_pRegistry;
-		std::map<Glory::UUID, size_t> m_ComponentTypes;
-		std::map<size_t, Glory::UUID> m_TypeToUUID;
+		std::map<Glory::UUID, uint32_t> m_ComponentTypes;
+		std::map<uint32_t, Glory::UUID> m_TypeToUUID;
 		std::vector<Glory::UUID> m_ComponentOrder;
 	};
 }
